clean code告警清理——类型转换引起的整数回绕
# message auto-generated for no-merge-commit merge:
merge clean-code-fix-int-cast into master

clean code告警清理——类型转换引起的整数回绕

Created-by: msprofDeveloper

Match-id-51b6295c28b3d62addd880192faaf5660350f969

+++ analysis/csrc/domain/services/modeling/batch_id/batch_id.cpp
@@ -74,7 +74,7 @@ int ModelingCalibrateBatchId(HalUniData **task, uint32_t taskNum, int cursor, ui
     return cursor;
 }
 
-void ModelingComputeBatchIdBinary(HalUniData **task, uint32_t taskNum, HalUniData **flip, uint32_t flipNum)
+void ModelingComputeBatchIdBinary(HalUniData **task, uint32_t taskNum, HalUniData **flip, uint16_t flipNum)
 {
     if (task == nullptr || taskNum == 0 || flip == nullptr || flipNum == 0) {
         ERROR("input error!");
